requirements.txt
aio-pika==9.6.2
aiormq==6.9.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
boto3==1.42.88
botocore==1.42.88
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.5
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
jmespath==1.1.0
multidict==6.7.1
pamqp==3.3.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
yarl==1.23.0
